+++ src/main/java/net/pedroksl/advanced_ae/common/entities/QuantumCrafterEntity.java
@@ -233,11 +233,11 @@ public class QuantumCrafterEntity extends AENetworkPowerBlockEntity
                 IPatternDetails details = PatternDetailsHelper.decodePattern(is, this.getLevel());
                 if (details instanceof AECraftingPattern craftPattern) {
                     if (craftingJobs.get(x) != null) {
-                        if (craftingJobs.get(x).pattern == craftPattern) {
+                        if (craftingJobs.get(x).pattern == null) {
+                            craftingJobs.get(x).setPattern(craftPattern);
                             setInvalidPattern(x, craftingJobs.get(x).consumesDurability);
                             continue;
-                        } else if (craftingJobs.get(x).pattern == null) {
-                            craftingJobs.get(x).setPattern(craftPattern);
+                        } else if (craftingJobs.get(x).pattern.equals(craftPattern)) {
                             setInvalidPattern(x, craftingJobs.get(x).consumesDurability);
                             continue;
                         }
@@ -385,7 +385,7 @@ public class QuantumCrafterEntity extends AENetworkPowerBlockEntity
     private boolean hasAvailableOutputStorage(CraftingJob job) {
         for (var output : job.pattern.getOutputs()) {
             if (output.what() instanceof AEItemKey key) {
-                var stack = key.toStack();
+                var stack = key.toStack((int) output.amount());
                 for (var x = 0; x < this.outputInv.size(); x++) {
                     stack = this.outputInv.insertItem(x, stack, true);
                     if (stack.isEmpty()) {
@@ -1148,7 +1148,7 @@ public class QuantumCrafterEntity extends AENetworkPowerBlockEntity
                 var in = input.getPossibleInputs()[0];
 
                 if (in.what() instanceof AEItemKey inKey) {
-                    var inStack = inKey.toStack();
+                    var inStack = inKey.toStack((int) in.amount());
 
                     if (!this.keysThatConsumeDurability.containsKey(inKey)) {
                         ItemStack remainingStack = findMatchingRemainingItem(in);
@@ -1302,7 +1302,7 @@ public class QuantumCrafterEntity extends AENetworkPowerBlockEntity
             ItemStack inStack = inKey.toStack();
 
             for (var output : pattern.getOutputs()) {
-                ItemStack outStack = ((AEItemKey) output.what()).toStack();
+                ItemStack outStack = ((AEItemKey) output.what()).toStack((int) output.amount());
 
                 if (inStack.getOrCreateTag().equals(outStack.getOrCreateTag()) || inStack.is(outStack.getItem())) {
                     return outStack;
@@ -1364,7 +1364,7 @@ public class QuantumCrafterEntity extends AENetworkPowerBlockEntity
                 if (input == null) {
                     craftingInput.add(ItemStack.EMPTY);
                 } else if (input.what() instanceof AEItemKey key) {
-                    craftingInput.add(key.toStack());
+                    craftingInput.add(key.toStack((int) input.amount()));
                 }
                 if (pattern.canSubstituteFluids() && pattern.getValidFluid(x) != null) {
                     bucketsToRemove++;